Move out vg kmers
